pyproject.toml
[project]
name = "phantomfetch"
version = "0.1.2"
description = "High-performance agentic web scraping library combining curl-cffi speed with Playwright browser capabilities"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "CosmicBull", email = "cosmicbull@example.com"}
]
license = {text = "MIT"}
keywords = [
    "web-scraping",
    "playwright",
    "curl-cffi",
    "async",
    "browser-automation",
    "http-client",
    "agentic",
    "anti-detection"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO",
]
dependencies = [
    "curl-cffi>=0.13.0",
    "msgspec>=0.20.0",
    "playwright>=1.56.0",
    "rusticsoup>=0.3.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.38.0",
    "loguru>=0.7.3",
]

[project.urls]
Homepage = "https://github.com/iristech-systems/PhantomFetch"
Documentation = "https://github.com/iristech-systems/PhantomFetch#readme"
Repository = "https://github.com/iristech-systems/PhantomFetch"
Issues = "https://github.com/iristech-systems/PhantomFetch/issues"
Changelog = "https://github.com/iristech-systems/PhantomFetch/blob/main/CHANGELOG.md"

[project.scripts]
phantomfetch = "phantomfetch:main"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "myst-parser>=4.0.1",
    "opentelemetry-sdk>=1.38.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "sphinx>=8.2.3",
    "sphinx-wagtail-theme>=6.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["notebooks"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "PL", "RUF"]
ignore = ["PLR0913", "PLR0912", "E501", "PLR0915", "PLR2004", "PLR0911", "PLC0415"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["rusticsoup.*", "uvloop.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*", "examples.*", "notebooks.*"]
disallow_untyped_defs = false
